+++ src/Utils/ObjectBeanPropertyDescriptor.php
@@ -19,6 +19,8 @@ use Zend\Code\Generator\ParameterGenerator;
  */
 class ObjectBeanPropertyDescriptor extends AbstractBeanPropertyDescriptor
 {
+    use ForeignKeyAnalyzerTrait;
+
     /**
      * @var ForeignKeyConstraint
      */
@@ -28,14 +30,6 @@ class ObjectBeanPropertyDescriptor extends AbstractBeanPropertyDescriptor
      */
     private $beanNamespace;
 
-    /**
-     * @var Annotations[]
-     */
-    private $annotations;
-    /**
-     * @var Column[]
-     */
-    private $localColumns;
     /**
      * @var AnnotationParser
      */
@@ -100,19 +94,6 @@ class ObjectBeanPropertyDescriptor extends AbstractBeanPropertyDescriptor
         }
 
         return false;
-    }
-
-    /**
-     * @return Column[]
-     */
-    private function getLocalColumns(): array
-    {
-        if ($this->localColumns === null) {
-            $localColumnNames = $this->foreignKey->getUnquotedLocalColumns();
-
-            $this->localColumns = array_map([$this->table, 'getColumn'], $localColumnNames);
-        }
-        return $this->localColumns;
     }
 
     /**
@@ -264,19 +245,4 @@ class ObjectBeanPropertyDescriptor extends AbstractBeanPropertyDescriptor
         return false;
     }
 
-    /**
-     * @return Annotations[]
-     */
-    private function getAnnotations(): array
-    {
-        if ($this->annotations === null) {
-            $this->annotations = [];
-
-            // Are all columns nullable?
-            foreach ($this->getLocalColumns() as $column) {
-                $this->annotations[] = $this->annotationParser->getColumnAnnotations($column, $this->table);
-            }
-        }
-        return $this->annotations;
-    }
 }