+++ tests/TDBMAbstractServiceTest.php
@@ -332,7 +332,7 @@ abstract class TDBMAbstractServiceTest extends TestCase
             ->column('id')->integer()->primaryKey()->autoIncrement()->comment('@Autoincrement')
             ->column('file')->blob()
             ->column('md5')->string()->null()->comment("@ProtectedGetter\n@ProtectedSetter")
-            ->column('article_id')->references('article')->null()->comment("@ProtectedGetter\n@ProtectedSetter");
+            ->column('article_id')->references('article')->null()->comment("@ProtectedGetter\n@ProtectedSetter\n@ProtectedOneToMany");
 
         $toSchema->getTable('users')
             ->addUniqueIndex([$connection->quoteIdentifier('login')], 'users_login_idx')