+++ CHANGELOG.md
@@ -7,6 +7,7 @@ New features:
 - \#95: Table and column annotations are now parsed using doctrine/annotations lexer/parser
 - \#97: Generated DAOs and beans are now purged automatically. When a table is removed, the matching generated beans and daos will be removed too.
 - \#116: New CodeGeneratorListernerInterface allows third party library to alter generated beans and DAOs on the fly (for the power users!)
+- \#125: New @ProtectedGetter, @ProtectedSetter and @ProtectedOneToMany annotations (to be used in the DB column comments) enable generating beans with protected getters and setters
 
 5.0
 ===